refactor: move validator accessor to EpochManagerHandle (#7655)
Part of https://github.com/near/nearcore/issues/6910, depends on #7652

Must be reviewed per commit.
